MAINT: move test to util, handle buggy NumPy versions
